Add configured settings support and ignoring body nodes
Added the [global] block for the jikken configuration. The first
setting added here is a "continue on failure" setting so if a test
fails to run, the tests will continue to execute.

Also added support for the RI file tag to prune root nodes from the JSON.
This support also works for comparison based tests.